separate compilation and runtime vm
