+++ python/tvm/autotvm/task/relay_integration.py
@@ -39,7 +39,6 @@ def _lower(mod,
     """
     # pylint: disable=import-outside-toplevel
     from tvm import relay
-    from tvm import runtime
     from tvm.relay.backend import graph_runtime_codegen
 
     if hasattr(target, 'device_name') and target.device_name == "vta":
@@ -50,7 +49,7 @@ def _lower(mod,
                 grc = graph_runtime_codegen.GraphRuntimeCodegen(None, target)
                 grc.codegen(mod["main"])
     # default case
-    compiler = runtime.vm.VMCompiler()
+    compiler = relay.vm.VMCompiler()
     if params:
         compiler.set_params(params)
     compiler.lower(mod, target=target)
